requirements.txt
absl-py==2.1.0
aiofiles==23.2.1
alembic==1.13.2
#ane_transformers==0.1.1
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
anyio==4.4.0
asgiref==3.8.1
asteroid-filterbanks==0.4.0
audioread==3.0.1
av==12.3.0
bitsandbytes==0.37.0
cattrs==24.1.1
certifi
cffi==1.17.1
chardet==5.2.0
charset-normalizer
click==8.1.7
coloredlogs==15.0.1
colorlog==6.8.2
coremltools==7.0b2
cryptography==42.0.8
ctranslate2==4.4.0
Cython>=0.29
datasets==3.0.0
dill
Django==4.2.15
django-auditlog==3.0.0
django-filter==23.5
django-js-asset==2.2.0
django-mptt==0.14.0
djangorestframework==3.15.1
drf-excel==2.4.0
drf-flex-fields==1.0.2
einops==0.8.0
et-xmlfile==1.1.0
evaluate==0.4.3
fastapi==0.114.2
faster-whisper==1.0.3
ffmpy==0.4.0
filelock
flatbuffers==24.3.25
fsspec
gmpy2
gradio==4.44.0
gradio_client==1.3.0
grpcio==1.66.1
h11==0.14.0
httpcore==1.0.5
httpx==0.27.2
huggingface_hub
humanfriendly==10.0
HyperPyYAML==1.2.2
idna
jiwer==3.0.4
joblib==1.4.2
julius==0.2.7
lazy_loader==0.4
ldap3==2.9.1
librosa==0.10.2.post1
lightning==2.4.0
lightning-utilities==0.11.7
llvmlite==0.43.0
Mako==1.3.5
Markdown==3.7
markdown-it-py==3.0.0
mdurl==0.1.2
more-itertools==10.5.0
mpmath
msgpack==1.1.0
NEMO==6.0.3
nemo-toolkit==1.21.0
networkx
ninja==1.11.1.1
numba==0.60.0
numexpr
numpy==1.23.5
omegaconf==2.3.0
onnx==1.16.2
onnxruntime==1.19.2
openai-whisper
openpyxl==3.1.5
optimum-quanto==0.2.4
optuna==4.0.0
orjson==3.10.7
packaging
pandas
pillow==10.3.0
pooch==1.8.2
praat-textgrids==1.4.0
primePy==1.3
#protobuf==3.20.1
psutil
pyaml==24.7.0
pyannote.audio==3.3.2
pyannote.core==5.0.0
pyannote.database==5.1.0
pyannote.metrics==3.2.1
pyannote.pipeline==3.0.1
pyarrow
pyasn1==0.6.1
PyAudio==0.2.14
pycparser==2.22
pydantic==2.9.1
pydantic_core==2.23.3
pydub==0.25.1
pymodbus==3.3.2
pyparsing
PySocks
python-dateutil
python-multipart==0.0.9
pytorch-lightning==2.4.0
pytorch-metric-learning==2.6.0
pytz
PyYAML
rapidfuzz==3.9.7
regex
requests
rich==13.8.1
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.8
ruff==0.6.5
sacremoses==0.1.1
safetensors
scikit-learn==1.5.2
scipy==1.14.1
semantic-version==2.10.0
semver==3.0.2
sentencepiece==0.2.0
shellingham==1.5.4
six
sniffio==1.3.1
sortedcontainers==2.4.0
sounddevice==0.5.0
soundfile==0.12.1
soxr==0.5.0.post1
speechbrain==1.0.1
SpeechRecognition
SQLAlchemy==2.0.34
sqlparse==0.5.1
starlette==0.38.5
sympy
tabulate==0.9.0
tensorboard==2.17.1
tensorboard-data-server==0.7.2
tensorboardX==2.6.2.2
text-unidecode==1.3
threadpoolctl==3.5.0
tiktoken==0.7.0
tokenizers
tomlkit==0.12.0
torch==2.4.1
torch-audiomentations==0.11.1
torch-pitch-shift==1.2.4
torchaudio==2.4.1
torchmetrics==1.4.1
torchvision==0.19.1
tqdm
transformers
typer==0.12.5
typing_extensions
tzdata
unicodedata2
urllib3
uvicorn==0.30.6
webrtcvad==2.0.10
websockets==12.0
Werkzeug==3.0.4
wget==3.2
wrapt==1.16.0
xxhash
